+++ database/migrations/2025_08_19_165929_create_espelhos_table.php
@@ -18,10 +18,10 @@ return new class extends Migration {
             $table->foreignId('cidade_id')->unique()->constrained('politica_cidades')->cascadeOnDelete();
             $table->string('presidente_local')->nullable();
             $table->string('indicacao_bispo')->nullable();
-            $table->integer('filiados_republicanos')->nullable();
+            $table->unsignedInteger('filiados_republicanos')->nullable(); // Corrigido
             $table->string('prefeito_atual_nome')->nullable();
             $table->string('prefeito_atual_partido')->nullable();
-            $table->integer('prefeito_atual_votos')->nullable();
+            $table->unsignedBigInteger('prefeito_atual_votos')->nullable(); // Corrigido
             $table->text('observacoes')->nullable();
             $table->timestamps();
         });
